% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_contrasts.R
\name{as_is}
\alias{as_is}
\title{Use contrast matrix as-is}
\usage{
as_is(x)
}
\arguments{
\item{x}{Contrast matrix or contrast function}
}
\value{
x
}
\description{
The default behavior for setting contrasts with a reference level is to move
the reference level to be the first level. However, there are some functions
that do not natively specify that already. Consider \code{contr.treatment} and
\code{contr.SAS}, which provide the same treatment-coded matrices BUT the former
moves the reference level to the first level while the latter sets the
reference level to the last level. The contrast between the two functions is
thus neutralized in this package. If we have reason to avoid this behavior,
especially for manually created matrices for niche situations, we can
suppress the behavior by wrapping the matrix/contrast function in \code{as_is}.
}
\details{
By itself, this function does nothing more than return its input.
For the purposes of this package though, some functions will check to see if
a contrast matrix is wrapped in \code{as_is}.
}
\examples{

# These will return the same thing
enlist_contrasts(mtcars, carb ~ contr.treatment)
enlist_contrasts(mtcars, carb ~ contr.SAS)

# This will suppress moving the reference level
enlist_contrasts(mtcars, carb ~ as_is(contr.SAS))

# Particularly helpful for external matrices
# Imagine this is some other more complicated matrix
my_contrasts <- contr.SAS(6)

# This is the same as as_is(contr.SAS)
enlist_contrasts(mtcars, carb ~ as_is(my_contrasts))
}
