% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_contrasts.R
\name{.add_namespace}
\alias{.add_namespace}
\title{Lookup namespace of contrast scheme function}
\usage{
.add_namespace(scheme_names)
}
\arguments{
\item{scheme_names}{Character vector}
}
\value{
character vector of updated function names with namespaces
}
\description{
Given the name of a contrast scheme (ie the function name that creates
the contrast matrix), lookup which namespace it belongs to and add it
to the string. Used to report which packages are used in the glimpse table,
e.g., stats, contrastable, bayesTestR
}
