% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_contrast.R
\name{.bundle_params}
\alias{.bundle_params}
\title{Extract parameters from dots}
\usage{
.bundle_params(factor_col, ...)
}
\arguments{
\item{factor_col}{Factor to set contrasts to}

\item{...}{Other arguments passed by the user}
}
\value{
Function call parameters as a list
}
\description{
If there are any other arguments passed to a contrast coding function by
the user, ensure they're bundled together correctly. This is especially
important for setting the number of levels correctly.
}
\keyword{internal}
