% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_where.R
\name{.cols_where}
\alias{.cols_where}
\title{Get columns where}
\usage{
.cols_where(model_data, fx, use_names = FALSE, return_names = FALSE)
}
\arguments{
\item{model_data}{Model data}

\item{fx}{Function to apply, must be something that returns a logical value.
Usually either \code{is.factor} or \code{is.ordered}}

\item{use_names}{Whether the resulting vector should be named}

\item{return_names}{Whether names (where the fx returns TRUE) should be
returned instead of a logical vector. Overwrites use.names.}
}
\value{
Optionally named logical vector or character vector depending on
values of \code{use_names} and \code{return_names}
}
\description{
Helper to avoid the use of tidyselect and dplyr::select, returns either a
logical vector (optionally named) or a character vector of which columns
satisfy the given function
}
