% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{.contrasts_to_hypotheses}
\alias{.contrasts_to_hypotheses}
\title{Convert contrast matrix to hypothesis matrix}
\usage{
.contrasts_to_hypotheses(contrast_matrix)
}
\arguments{
\item{contrast_matrix}{Contrast matrix}
}
\value{
Hypothesis matrix
}
\description{
Convert contrast matrix to hypothesis matrix
}
\examples{

contrastable:::.contrasts_to_hypotheses(scaled_sum_code(5))
}
\keyword{internal}
