% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{.convert_matrix}
\alias{.convert_matrix}
\title{Convert between hypothesis and contrast matrices}
\usage{
.convert_matrix(m)
}
\arguments{
\item{m}{matrix}
}
\value{
Hypothesis matrix if passed contrasts, contrast matrix if passed
hypotheses
}
\description{
Quick debugging helper that runs the correct matrix conversion based on
the dimensions of the matrix
}
\keyword{internal}
