% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels_as_needed.R
\name{.droplevels_as_needed}
\alias{.droplevels_as_needed}
\title{Drop missing levels}
\usage{
.droplevels_as_needed(model_data, lhs_variables, verbose = TRUE)
}
\arguments{
\item{model_data}{Model data}

\item{lhs_variables}{variables specified for contrast coding from formulas}

\item{verbose}{Should messages be sent? Defaults to TRUE}
}
\value{
model_data where existing factor columns have had any missing levels
removed
}
\description{
Missing levels can cause unexpected behavior when fitting a model after
setting contrasts. This function removes missing levels and informs the user
of how many levels were dropped from each factor.
}
