% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_formulas.R
\name{.expand_formulas}
\alias{.expand_formulas}
\title{Expand contrast formulas}
\usage{
.expand_formulas(formulas, data)
}
\arguments{
\item{formulas}{List of formulas}

\item{data}{Dataframe to evaluate names in}
}
\value{
Named list of formulas
}
\description{
Uses \verb{\{tidyselect]} functionality to expand the left hand side of a formula
into multiple formulas. See examples of \code{enlist_contrasts} for examples.
}
