% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_contrast.R
\name{.get_dimnames}
\alias{.get_dimnames}
\title{Get dimnames of contrasts from factor}
\usage{
.get_dimnames(factor_col)
}
\arguments{
\item{factor_col}{Factor to extract contrasts from}
}
\value{
List of rownames and column names
}
\description{
Given a factor, extract the row and column names of the contrasts. If they're
not set, then use default values. Also helps to catch invalid usage of
polynomial contrasts.
}
\keyword{internal}
