% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_contrasts.R
\name{.get_reference_level}
\alias{.get_reference_level}
\title{Get reference level of contrast matrix}
\usage{
.get_reference_level(contrast_matrix)
}
\arguments{
\item{contrast_matrix}{Contrast matrix, if NULL will throw an error}
}
\description{
Compute the reference level manually if the information isn't available
via parameters
}
\seealso{
\code{\link[=.get_reference_levels]{.get_reference_levels()}}
}
\keyword{internal}
