% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{.get_reserved_operator}
\alias{.get_reserved_operator}
\title{Check for reserved operators}
\usage{
.get_reserved_operator(node)
}
\arguments{
\item{node}{A symbol, extracted from a formula}
}
\value{
If \code{node} is a reserved operator, then return the operator as a
string. Otherwise return the string "none".
}
\description{
A helper for \code{\link[=.make_parameters]{.make_parameters()}}, which takes a symbol and checks
whether it corresponds to one of the reserved operators for the package
syntax.
}
\keyword{internal}
