% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_contrasts.R
\name{.get_scheme_labels}
\alias{.get_scheme_labels}
\title{Get contrast scheme labels for glimpse table}
\usage{
.get_scheme_labels(list_params, formulas)
}
\arguments{
\item{list_params}{List of params, see \code{\link[=.make_parameters]{.make_parameters()}}}

\item{formulas}{Formulas passed by the user}
}
\value{
Character vector
}
\description{
Looks up contrast functions from formulas and appends the necessary
namespaces
}
\keyword{internal}
