% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{.hypotheses_to_contrasts}
\alias{.hypotheses_to_contrasts}
\title{Convert hypothesis matrix to contrast matrix}
\usage{
.hypotheses_to_contrasts(hypothesis_matrix)
}
\arguments{
\item{hypothesis_matrix}{Hypothesis matrix}
}
\value{
Hypothesis matrix
}
\description{
Convert hypothesis matrix to contrast matrix
}
\examples{

# Treatment coding
contrastable:::.hypotheses_to_contrasts(matrix(c(1,0,-1,1), nrow=2))
}
\keyword{internal}
