% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts_polynomial.R
\name{.is_polynomial_scheme}
\alias{.is_polynomial_scheme}
\title{Check if a scheme is a polynomial scheme}
\usage{
.is_polynomial_scheme(scheme)
}
\arguments{
\item{scheme}{Contrast coding function or string of a function name}
}
\value{
\code{TRUE} if polynomial scheme, \code{FALSE} otherwise
}
\description{
Helper to check if a contrast coding function is one of the polynomial
schemes
}
\keyword{internal}
