% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{.parse_formula}
\alias{.parse_formula}
\title{Parse contrast formula}
\usage{
.parse_formula(raw_formula)
}
\arguments{
\item{raw_formula}{Raw formula passed by user}
}
\value{
A list of parameters to use for a use_contrasts call
}
\description{
Takes a formula and extracts the parameters for \code{use_contrasts} if it is
a valid formula. Validity checked via regexes, parameters extracted via
recursively parsing the abstract syntax tree.
}
