% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_matrix.R
\name{.postprocess_matrix}
\alias{.postprocess_matrix}
\title{Postprocess contrast matrices}
\usage{
.postprocess_matrix(
  new_contrasts,
  code_by,
  reference_level,
  set_intercept,
  drop_trends,
  matrix_labels,
  labels,
  as_is,
  dots
)
}
\arguments{
\item{new_contrasts}{Contrast matrix passed from \code{\link[=use_contrasts]{use_contrasts()}}}

\item{code_by}{What to code with, passed from \code{\link[=use_contrasts]{use_contrasts()}}}

\item{reference_level}{Reference level specification}

\item{set_intercept}{Intercept specification}

\item{drop_trends}{Any trends to drop}

\item{matrix_labels}{Labels from the matrix dimension names}

\item{labels}{User-specified labels}

\item{as_is}{Whether to use matrix as-is}

\item{dots}{Any additional arguments passed with code_by}
}
\value{
Contrast matrix
}
\description{
When using \code{\link[=use_contrasts]{use_contrasts()}} with a function or matrix, there are shared
postprocessing steps for manipulating the contrasts to set the reference
level, intercept, and labels. If as_is is TRUE, then these post processing
steps aren't included.
}
\seealso{
\code{\link[=use_contrasts]{use_contrasts()}}
}
\keyword{internal}
