% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{.process_code_by}
\alias{.process_code_by}
\title{Process code_by}
\usage{
.process_code_by(formula, params, env, verbose)
}
\arguments{
\item{formula}{Formula used to set contrast}

\item{params}{Parameter list}

\item{env}{Not used}

\item{verbose}{Whether to throw the warning about invalid \code{-} usage}
}
\value{
Modified parameter list with updated \code{code_by}
}
\description{
Handles the \code{code_by} parameter and checks to make sure whether we can
safely drop trends or not. Also handles any usage of \code{\link[=I]{I()}}
}
\keyword{internal}
