% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist_contrasts.R
\name{.process_contrasts}
\alias{.process_contrasts}
\title{Pass arguments to contrast code}
\usage{
.process_contrasts(model_data, raw_formula)
}
\arguments{
\item{model_data}{Data frame with factor column}

\item{raw_formula}{Raw formula}
}
\value{
A contrast matrix
}
\description{
Processes a formula and any arguments for the contrast matrix and sets up
the use_contrasts call
}
