% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{.process_factor_col}
\alias{.process_factor_col}
\title{Process factor column}
\usage{
.process_factor_col(cur_expr, params, env, verbose)
}
\arguments{
\item{cur_expr}{Current expression, see \code{\link[=.make_parameters]{.make_parameters()}}}

\item{params}{Parameter accumulator}

\item{env}{Environment for the formula, on the first recursion this will be
pulled from \code{formula} and then passed to subsequent recursions. Needed to
check whether a function is actually a function.}

\item{verbose}{Logical, default \code{TRUE}, whether to show drop_trends warning
if used incorrectly}
}
\value{
Parameter list with \code{factor_col} updated
}
\description{
This is usually the first thing to be processed in a formula like
\code{varname ~ x}; ensures that \code{varname} is treated as a factor vector
}
\keyword{internal}
