% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_labels.R
\name{.reset_comparison_labels}
\alias{.reset_comparison_labels}
\title{Reset comparison labels of matrix to defaults}
\usage{
.reset_comparison_labels(contr_mat, coding_fx = NULL)
}
\arguments{
\item{contr_mat}{Contrast matrix}

\item{coding_fx}{Function name as a string}
}
\value{
Matrix with column names changed if necessary
}
\description{
Given a contrast matrix and a coding function used to generate it, check
whether we have default labels implemented. If so, use them if the matrix
doesn't have unique ones. If no function is provided, just use what the
matrix has or use numeric indices
}
\keyword{internal}
