% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_contrast.R
\name{.set_intercept}
\alias{.set_intercept}
\title{Set intercept for contrast matrix}
\usage{
.set_intercept(contrast_matrix, intercept_level)
}
\arguments{
\item{contrast_matrix}{Contrast matrix to use}

\item{intercept_level}{Level to use as intercept, must be present in the
row names of the contrast matrix}
}
\value{
Manipulated contrast matrix
}
\description{
Given a contrast matrix and a desired level, change the intercept via
manipulating underlying hypothesis matrix.
}
\keyword{internal}
