% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{.set_param}
\alias{.set_param}
\title{Process and set parameter}
\usage{
.set_param(cur_expr, params, env, which_param)
}
\arguments{
\item{cur_expr}{Current expression, a formula or list representation thereof}

\item{params}{Named list of parameters}

\item{env}{Environment to evaluate expressions in \code{cur_expr} in}

\item{which_param}{Which parameter to set, a string, see \code{make_parameters}
for usage}
}
\value{
\code{params}
}
\description{
Unpacks the given expression to set the parameter specified by \code{which_param}
to \code{params}. Continues recursively setting parameters via \code{make_parameters}.
}
