% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_reference_level.R
\name{.switch_reference_if_needed}
\alias{.switch_reference_if_needed}
\title{Check if reference switching is necessary}
\usage{
.switch_reference_if_needed(
  contrast_matrix,
  new_reference_label = NA,
  new_reference_index
)
}
\arguments{
\item{contrast_matrix}{Contrast matrix}

\item{new_reference_label}{Name of the level to use}

\item{new_reference_index}{Index of the level to use}
}
\value{
Contrast matrix with reference level swapped if needed
}
\description{
Check if the contrast matrix is able to switch reference levels, and if it
is requested, do so
}
\keyword{internal}
