% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_reference_level.R
\name{.switch_reference_level}
\alias{.switch_reference_level}
\title{Switch reference level in contrast matrix}
\usage{
.switch_reference_level(contrast_matrix, old_reference, new_reference)
}
\arguments{
\item{contrast_matrix}{An unlabeled contrast matrix}

\item{old_reference}{The previously specified reference level, usually the
default for the scheme found by \code{.get_reference_level}}

\item{new_reference}{The index of the row to use as the new reference level}
}
\value{
A matrix with the reordered rows and correct values for the reference
level
}
\description{
Reorders the rows of a contrast matrix to properly set the reference level
}
\keyword{internal}
