% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_orthogonal.R
\name{is_orthogonal}
\alias{is_orthogonal}
\title{Check for orthogonality}
\usage{
is_orthogonal(.contrasts, USE.NAMES = FALSE)
}
\arguments{
\item{.contrasts}{Contrast matrix or list of contrast matrices}

\item{USE.NAMES}{Logical, whether vector should be named}
}
\value{
Logical vector, will retain names of a passed list
}
\description{
Given a contrast matrix or list of contrast matrices (eg from \code{enlist_contrasts}),
return a logical vector of whether each contrast is orthogonal or not.
}
