% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_contrasts.R
\name{print_contrasts}
\alias{print_contrasts}
\title{Print contrasts}
\usage{
print_contrasts(x, specific_vars)
}
\arguments{
\item{x}{Dataframe containing factor columns}

\item{specific_vars}{Character vector of specific factor columns to print
contrasts for. Use to only print a certain subset of variables.}
}
\value{
Nothing
}
\description{
Utility to print all contrast matrices for a given dataframe.
Displays fractional values with \code{\link[MASS:fractions]{MASS::fractions()}}.
Prints nothing if there are no factor columns in the dataframe.
}
\examples{

print_contrasts(data.frame(x = gl(3, 1), y = gl(3, 1, ordered = TRUE)))
}
