% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_contrast_function.R
\name{use_contrast_function}
\alias{use_contrast_function}
\title{Set contrast coding with function}
\usage{
use_contrast_function(
  factor_col,
  code_by,
  reference_level = NA,
  set_intercept = NA,
  drop_trends = NA,
  as_is = FALSE,
  ...
)
}
\arguments{
\item{factor_col}{Factor column}

\item{code_by}{A function that generates contrast matrices}

\item{reference_level}{Label of the desired reference level, will default to
the first label in dimnames of the original contrasts}

\item{set_intercept}{Label of the desired intercept level, will default to
scheme's default (typically grand mean, unless using contr.treatment, which does
the reference level)}

\item{drop_trends}{Which trends to drop from the final matrix, should only
be used with \code{contr.poly} to remove higher order trends}

\item{as_is}{Logical, default FALSE, whether to suppress auto switching of
the reference level to the first level if not specified}

\item{...}{Other arguments to pass to \code{coding_fx}}
}
\description{
Set contrast coding with function
}
