% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_contrast_matrix.R
\name{use_contrast_matrix}
\alias{use_contrast_matrix}
\title{Manual code factors}
\usage{
use_contrast_matrix(factor_col, code_by = NA, as_is = FALSE, ...)
}
\arguments{
\item{factor_col}{The factor column to use, eg data$gender}

\item{code_by}{The matrix you want to use, obligatory if n factors > 2}

\item{as_is}{Whether to use the factor levels as labels}

\item{...}{Not used}
}
\value{
A matrix for contrast coding schemes
}
\description{
Use this when you want to manually specify the contrast matrix
}
