% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_contrast.R
\name{use_contrasts.array}
\alias{use_contrasts.array}
\title{Matrix handler for use_contrasts
Passing a matrix uses the implicit array class, so just pass it off to the
matrix method}
\usage{
\method{use_contrasts}{array}(
  factor_col,
  code_by = NA,
  reference_level = NA,
  set_intercept = NA,
  drop_trends = NA,
  labels = NULL,
  as_is = FALSE,
  ...
)
}
\arguments{
\item{factor_col}{A factor vector, eg from df$factorVarName}

\item{code_by}{A matrix to be used as the contrast matrix}

\item{reference_level}{The level to use as the reference level, default NA}

\item{set_intercept}{The intercept to use, default NA}

\item{drop_trends}{The trends to drop, default NA}

\item{labels}{A vector of labels to apply to the matrix column names, default}

\item{as_is}{Logical, default FALSE, whether to leave the resulting matrix}

\item{...}{Additional arguments to be passed to \code{code_by()}}
}
\value{
A contrast coding matrix with labels and proper reference level
}
\description{
Matrix handler for use_contrasts
Passing a matrix uses the implicit array class, so just pass it off to the
matrix method
}
\examples{

# Works the same as use_contrasts.matrix
contrast_matrix <- sum_code(4)
use_contrasts.array(gl(4,1), contrast_matrix)
}
