% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{.set_param}
\alias{.set_param}
\title{Process and set parameter}
\usage{
.set_param(cur_expr, params, env, which_param, verbose)
}
\arguments{
\item{cur_expr}{Current expression, a formula or list representation thereof}

\item{params}{Parameter accumulator}

\item{env}{Environment for the formula, on the first recursion this will be
pulled from \code{formula} and then passed to subsequent recursions. Needed to
check whether a function is actually a function.}

\item{which_param}{Which parameter to set, a string
for usage}

\item{verbose}{Logical, default \code{TRUE}, whether to show drop_trends warning
if used incorrectly}
}
\value{
\code{params}
}
\description{
Unpacks the given expression to set the parameter specified by \code{which_param}
to \code{params}. Continues recursively setting parameters via
\code{\link[=.make_parameters]{.make_parameters()}}.
}
\seealso{
\code{\link[=.make_parameters]{.make_parameters()}}
}
\keyword{internal}
