% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unordered.R
\name{is.unordered}
\alias{is.unordered}
\title{Check for unordered factor}
\usage{
is.unordered(x)
}
\arguments{
\item{x}{a vector of data}
}
\value{
TRUE if x is an unordered factor, FALSE if x is not a factor or is an
ordered factor
}
\description{
Helper to check if a factor is exclusively unordered. is.factor(x) is TRUE
when x is unordered OR ordered.
}
\examples{
is.unordered(gl(5,1)) # True
is.unordered(gl(5,1,ordered = TRUE)) # False
}
