% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_orthogonal.R
\name{is_orthogonal}
\alias{is_orthogonal}
\title{Check for orthogonality}
\usage{
is_orthogonal(contrast_matrices, USE.NAMES = FALSE)
}
\arguments{
\item{contrast_matrices}{Contrast matrix or list of contrast matrices}

\item{USE.NAMES}{Logical, whether vector should be named}
}
\value{
Logical vector, will retain names of a passed list
}
\description{
Given a contrast matrix or list of contrast matrices (eg from
\code{\link[=enlist_contrasts]{enlist_contrasts()}} ), return a logical vector of whether each contrast is
orthogonal or not.
}
\examples{

is_orthogonal(treatment_code(5)) # FALSE
is_orthogonal(helmert_code(5)) # TRUE
}
