% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_contrast.R
\name{use_contrasts}
\alias{use_contrasts}
\title{Contrast code factors}
\usage{
use_contrasts(factor_col, code_by = NA, labels = NULL, as_is = FALSE, ...)
}
\arguments{
\item{factor_col}{The factor column to use, eg data$gender}

\item{code_by}{Either a matrix or a function}

\item{labels}{Labels to use in the contrast matrix, must equal number of contrasts}

\item{as_is}{Logical, default FALSE, whether to suppress auto switching of
the reference level to the first level if not specified}

\item{...}{Additional arguments to be passed to use_contrast_function, specifically,
which level you want the reference level to be}
}
\value{
A contrast coding matrix with labels and proper reference level
}
\description{
Helper to do contrast coding. There are two options:
\itemize{
\item Manually specify a matrix for code_by (implements use_contrast_matrix). Reference level
is automatically set to the row that's always negative.
\item Specify a style of contrast coding as a function (implements use_contrast_function).
Label of the reference level should be specified in ...
}
}
